// lab2/project/frontend/src/types/posts.types.ts
export interface PostAuthor {
  id: string;
  avatarUrl: string;
  fullName: string;
  email: string;
}

export interface Post {
  id: string;
  title: string;
  description: string;
  imageUrl: string;
  author: PostAuthor;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostDto {
  title: string;
  description: string;
  imageUrl: string;
  authorId: string;
}